+++ src/vue/pages/Customers.vue
@@ -1,6 +1,14 @@
 <template>
   <div class="customers-page">
     <top-bar>
+      <template slot="main">
+        <multi-select-field
+          v-if="selectionOptions.length"
+          @selected="emitSelectedBalances"
+          :options="selectionOptions"
+          select-all
+        />
+      </template>
       <template slot="extra">
         <button
           v-ripple
@@ -49,11 +57,14 @@
 <script>
 import TopBar from '@/vue/common/TopBar'
 import Drawer from '@/vue/common/Drawer'
-import { vueRoutes } from '@/vue-router/routes'
-
+import MultiSelectField from '@/vue/fields/MultiSelectField'
 import MassPaymentForm from '@/vue/forms/MassPaymentForm'
 import MassInvitationForm from '@/vue/forms/MassInvitationForm'
+
 import { Bus } from '@/js/helpers/event-bus'
+import { mapGetters, mapActions } from 'vuex'
+import { vueRoutes } from '@/vue-router/routes'
+import { vuexTypes } from '@/vuex'
 
 export default {
   name: 'customers-page',
@@ -63,6 +74,7 @@ export default {
     Drawer,
     MassPaymentForm,
     MassInvitationForm,
+    MultiSelectField,
   },
 
   data: _ => ({
@@ -70,7 +82,14 @@ export default {
     isPaymentDrawerShown: false,
     receivers: [],
     vueRoutes,
+    selectionOptions: [],
   }),
+
+  computed: {
+    ...mapGetters({
+      ownedAssets: vuexTypes.ownedAssets,
+    }),
+  },
 
   watch: {
     isPaymentDrawerShown (value) {
@@ -80,11 +99,17 @@ export default {
     },
   },
 
-  created () {
+  async created () {
+    await this.loadAssets()
     this.listen()
+    this.selectionOptions = this.getSelectionOptions()
   },
 
   methods: {
+    ...mapActions({
+      loadAssets: vuexTypes.LOAD_ASSETS,
+    }),
+
     emitUpdateList () {
       Bus.emit('customers:updateList')
     },
@@ -93,6 +118,19 @@ export default {
       Bus.on('customers:massIssue', payload => {
         this.receivers = ((payload || {}).receivers || [])
         this.isPaymentDrawerShown = true
+      })
+    },
+
+    emitSelectedBalances (values) {
+      Bus.emit('customers:showBalances', values)
+    },
+
+    getSelectionOptions () {
+      return this.ownedAssets.map(asset => {
+        return {
+          name: asset.name,
+          value: asset.code,
+        }
       })
     },
   },