+++ src/vue/fields/FileField.vue
@@ -168,7 +168,7 @@ export default {
   watch: {
     'value': function (value) {
       this.document = value
-      this.loadDocumentUrl(this.value)
+      if (value) this.loadDocumentUrl(value)
     },
   },