+++ CHANGELOG.md
@@ -7,7 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 ## [Unreleased]
+#### Added
+- Filter balances for customers
+
 #### Fixed
+- A bug when drop down go out of page boundaries
+- A bug with incorrect error message after enterring recipient email on transfer form
 - A bug when the user can not enter a value in input-field with type 'number' on Firefox
 
 #### Added