+++ CHANGELOG.md
@@ -9,14 +9,12 @@ for further information about branching and tagging conventions.
 ## [Unreleased]
 #### Added
 - Filter balances for customers
+- Promo codes
 
 #### Fixed
 - A bug when drop down go out of page boundaries
 - A bug with incorrect error message after enterring recipient email on transfer form
 - A bug when the user can not enter a value in input-field with type 'number' on Firefox
-
-#### Added
-- Promo codes
 
 ## [1.1.0] - 2019-10-12