+++ CHANGELOG.md
@@ -23,6 +23,9 @@ for further information about branching and tagging conventions.
 - A bug with incorrect error message after enterring recipient email on transfer form
 - A bug when the user can not enter a value in input-field with type 'number' on Firefox
 
+#### Changed
+- Renamed business to companies
+
 ## [1.1.0] - 2019-10-12
 
 ## [1.1.0-rc.3] - 2019-10-10